+++ throttle.ts
@@ -1,12 +1,12 @@
-import type { throttleInterface, returnFunctionInterface } from './type/throttle'
+import type { throttleInterface } from './type/throttle'
 
 /**
  * 节流
  * @param handler 逻辑函数
  * @param wait 执行逻辑的间隔
- * @returns returnFunctionInterface
+ * @returns Function
  */
-export const throttle: throttleInterface = (handler: Function, wait: number = 1000): returnFunctionInterface => {
+export const throttle: throttleInterface = function (handler: Function, wait: number = 1000): Function {
   let prev: number = Number(new Date)
   return function (): void {
     const now: number = Number(new Date)