+++ type/debounce.d.ts
@@ -1,5 +1,3 @@
-import { returnFunctionInterface } from './throttle'
-
 export interface debounceInterface {
-  (handle: Function, delay: number): returnFunctionInterface
+  (handle: Function, delay: number): Function
 }